Added fixes for some bugs
Added ability to recover from left over instances by deleting them.

Handling Http error responses in Restinator

Added Integration Tests
Added Concurrency Tests

Failure during creation now doesn't increment the instance manager

Cleaned up the JimController so it's easier to read

+++ dreidel-goyim/src/main/java/com/jive/qa/dreidel/goyim/views/DreidelView.java
@@ -1,8 +1,5 @@
 package com.jive.qa.dreidel.goyim.views;
 
-import java.io.IOException;
-import java.net.MalformedURLException;
-
 import javax.inject.Inject;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.POST;
@@ -11,8 +8,6 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
-import lombok.extern.slf4j.Slf4j;
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jive.qa.dreidel.goyim.controllers.InstanceManager;
@@ -20,11 +15,12 @@ import com.jive.qa.dreidel.goyim.controllers.JimController;
 import com.jive.qa.dreidel.goyim.exceptions.JimCreationException;
 import com.jive.qa.dreidel.goyim.exceptions.JimDestructionException;
 import com.jive.qa.dreidel.goyim.messages.ErrorMessage;
+import com.jive.qa.dreidel.goyim.messages.IdResponse;
 import com.jive.qa.dreidel.goyim.models.Instance;
+import com.jive.qa.dreidel.goyim.models.Network;
 import com.jive.qa.dreidel.goyim.service.BmSettings;
 
 @Path("/")
-@Slf4j
 public class DreidelView
 {
 
@@ -59,11 +55,38 @@ public class DreidelView
     try
     {
       jimController.createInstance(service, instance);
-      // TODO return id of the instance
-      return Response.status(Status.OK).build();
+
+      String address =
+          ((Network) instance.getNetworks().stream()
+              .filter(x -> x.getName().equals(settings.getPreferred()))
+              .toArray()[0]).getAddress();
+
+      return Response
+          .status(Status.OK)
+          .entity(
+              json.writeValueAsString(new IdResponse(instance.getInstance(), address)))
+          .build();
+
     }
     catch (JimCreationException e)
     {
+      // we need to remove the instance if there was a problem creating it.
+      instanceManager.removeInstance(instance.getInstance());
+      // return the error
+      // TODO exception mapper
+      return Response.status(Status.INTERNAL_SERVER_ERROR).entity(
+          json.writeValueAsString(new ErrorMessage(e.getClass().getSimpleName(), e.getMessage())))
+          .build();
+    }
+    catch (JimDestructionException e)
+    {
+      // don't delete the instance because there is something wrong with deleting it from jim.
+      // for now we are going to ignore this problem and just increment.
+      // no one will ever delete this instance because no one knows about it therefore it will just
+      // be a placeholder.
+
+      // return the error
+      // TODO exception mapper
       return Response.status(Status.INTERNAL_SERVER_ERROR).entity(
           json.writeValueAsString(new ErrorMessage(e.getClass().getSimpleName(), e.getMessage())))
           .build();
@@ -73,16 +96,31 @@ public class DreidelView
   @DELETE
   @Path("/{service}/{id}")
   public Response deleteServer(@PathParam("service") String service,
-      @PathParam("id") int id) throws MalformedURLException, IOException
+      @PathParam("id") int id) throws JsonProcessingException
   {
+    if (!jimController.serviceExists(service))
+    {
+      return Response.status(Status.NOT_FOUND).entity(
+          json.writeValueAsString(new ErrorMessage("NonExistantService", "The service "
+              + service + " does not exist."))).build();
+    }
+    if (!jimController.instanceExists(service, id, settings.getSite()))
+    {
+      return Response.status(Status.NOT_FOUND).entity(
+          json.writeValueAsString(new ErrorMessage("NonExistantInstance", "The Instance "
+              + id + " does not exist."))).build();
+    }
     try
     {
+      // first remove it from jim
       jimController.deleteInstance(service, id, settings.getSite());
+      // if there wasn't an exception remove it from the instance manager.
       instanceManager.removeInstance(id);
       return Response.status(Status.OK).build();
     }
     catch (JimDestructionException e)
     {
+      // TODO exception mapper
       return Response
           .status(Status.INTERNAL_SERVER_ERROR)
           .entity(
@@ -90,4 +128,5 @@ public class DreidelView
           .build();
     }
   }
+
 }
